fixed the test_api error
